# Bookstore-backend/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: My Jakarta EE API
  version: 1.0.0
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The book ID
        isbn:
          type: string
          description: The book ISBN
        author:
          type: string
          description: The book author
        title:
          type: string
          description: The book title
        year:
          type: integer
          format: int64
          description: The book year of publication
        publisher:
          type: string
          description: The book publisher
        domain:
          type: string
          description: The book domain
      required:
        - id
        - isbn
        - author
        - title
        - author
...
    Rating:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The rating ID
        user_id:
          type: integer
          format: int64
          description: The user ID
        book_id:
          type: integer
          format: int64
          description: The book ID
        date_added:
          type: string
          format: date-time
          description: The date the rating was added
        rating:
          type: integer
          format: int64
          description: The rating
        comment:
          type: string
          description: The comment added to the rating
      required:
        - id
        - user_id
        - book_id
        - date_added
        - rating
        - comment
    BooksComment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The object ID
        user_id:
          type: integer
          format: int64
          description: The user ID
        book_id:
          type: integer
          format: int64
          description: The book ID
        comment:
          type: string
          description: The comment added to the book
      required:
        - id
        - user_id
        - book_id
        - comment
    OffersComment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The object ID
        user_id:
          type: integer
          format: int64
          description: The user ID
        request_id:
          type: integer
          format: int64
          description: The request ID
        comment:
          type: string
          description: The comment added to the request
      required:
        - id
        - user_id
        - request_id
        - comment
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/book:
    get:
      tags:
        - book
      summary: Returns a list of books
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      tags:
        - book
      summary: Adds a new book to the database
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Book object to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: The added book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /api/book/name/{name}:
      get:
        tags:
          - book
        parameters:
              - in: path
                name: name
                schema:
                  type: string
                required: true
                description: Name of the book to retrieve
        summary: Returns a book by name
        responses:
            '200':
              description: A book
              content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/Book'
            '404':
              description: The book was not found
              content:
                  text/plain:
                    schema:
                      type: string
  /api/book/{id}:
    get:
      tags:
        - book
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: The book ID
      summary: Returns a book by ID
      responses:
        '200':
          description: A book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: The book was not found
          content:
            text/plain:
              schema:
                type: string
  /api/book/status/{value}:
      get:
        tags:
          - book
        parameters:
              - in: path
                name: value
                schema:
                  type: string
                required: true
                description: Status of the book to retrieve
        summary: Returns a list of books by status
        security:
          - bearerAuth: []
        responses:
            '200':
              description: A list of books
              content:
                  application/json:
                    schema:
                      type: array
                      items:
                        $ref: '#/components/schemas/Book'
            '404':
              description: The book was not found
              content:
                  text/plain:
                    schema:
                      type: string
  /api/book/author/{author}:
      get:
        tags:
          - book
        parameters:
              - in: path
                name: author
                schema:
                  type: string
                required: true
                description: Author of the book to retrieve
        summary: Returns a list of books by author
        security:
          - bearerAuth: []
        responses:
            '200':
              description: A list of books
              content:
                  application/json:
                    schema:
                      type: array
                      items:
                        $ref: '#/components/schemas/Book'
            '404':
              description: The book was not found
              content:
                  text/plain:
                    schema:
                      type: string
...
  /api/rating:
    get:
      tags:
        - rating
      summary: Returns a list of ratings
      responses:
        '200':
          description: A list of ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rating'
    post:
      tags:
        - rating
      summary: Adds a new rating to the database
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Rating object to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
      responses:
        '200':
          description: The added rating
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
  /api/rated/{userId}:
        get:
            tags:
            - rating
            parameters:
                - in: path
                  name: userId
                  schema:
                    type: integer
                    format: int64
                    required: true
                  description: User ID of the rating to retrieve
            summary: Returns a list of ratings by user ID
            responses:
                '200':
                  description: A list of ratings
                  content:
                    application/json:
                        schema:
                          type: array
                          items:
                            $ref: '#/components/schemas/Rating'
                '404':
                  description: The rating was not found
                  content:
                    text/plain:
                        schema:
                          type: string
  /api/rating/book/{book_id}:
      get:
        tags:
          - rating
        parameters:
              - in: path
                name: book_id
                schema:
                  type: integer
                  format: int64
                required: true
                description: Book ID of the rating to retrieve
        summary: Returns a list of ratings by book ID
        responses:
            '200':
              description: A list of ratings
              content:
                  application/json:
                    schema:
                      type: array
                      items:
                        $ref: '#/components/schemas/Rating'
            '404':
              description: The rating was not found
              content:
                  text/plain:
                    schema:
                      type: string
  /api/rating/user/{user_id}:
      get:
        tags:
          - rating
        parameters:
              - in: path
                name: user_id
                schema:
                  type: integer
                  format: int64
                required: true
                description: User ID of the rating to retrieve
        security:
          - bearerAuth: [ ]
        summary: Returns a list of ratings by user ID
        responses:
            '200':
              description: A list of ratings
              content:
                  application/json:
                    schema:
                      type: array
                      items:
                        $ref: '#/components/schemas/Rating'
            '404':
              description: The rating was not found
              content:
                  text/plain:
                    schema:
                      type: string
...
  /api/offerscomment:
    get:
      tags:
        - offers comment
      summary: Returns a list of offers comments
      responses:
        '200':
          description: A list of offers comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OffersComment'
    post:
      tags:
        - offers comment
      summary: Adds a new offer comment to the database
      requestBody:
        description: Offer comment to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OffersComment'
      responses:
        '200':
          description: The added comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffersComment'
  /api/offerscomment/user/{userId}:
    get:
      tags:
        - offers comment
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int64
          required: true
          description: User ID of the offers comment to retrieve
      summary: Returns a list of offers comments by user ID
      responses:
        '200':
          description: A list of offers comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OffersComment'
        '404':
          description: The offers comment was not found
          content:
            text/plain:
              schema:
                type: string
  /api/bookscomment/book/{bookId}:
    get:
      tags:
        - books comment
      parameters:
        - in: path
          name: bookId
          schema:
            type: integer
            format: int64
          required: true
          description: Book ID of the comment to retrieve
      summary: Returns a list of comments by book ID
      responses:
        '200':
          description: A list of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BooksComment'
        '404':
          description: The comment was not found
          content:
            text/plain:
              schema:
                type: string
  /api/offerscomment/offer/{offerId}:
    get:
      tags:
        - offers comment
      parameters:
        - in: path
          name: offerId
          schema:
            type: integer
            format: int64
          required: true
          description: Offer ID of the offers comment to retrieve
      summary: Returns a list of offers comments by offer ID
      responses:
        '200':
          description: A list of offers comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OffersComment'
        '404':
          description: The offers comment was not found
          content:
            text/plain:
              schema:
                type: string
...
  /api/book/user/{userId}:
    get:
      tags:
        - book
      summary: Returns a list of books by user id
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: ID of the user to retrieve books
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'